Use an int for the page index instead of a float

+++ engine/particle/src/particle.cpp
@@ -1173,11 +1173,11 @@ namespace dmParticle
             tile += start_tile;
             float* tex_coord = &tex_coords[tile << 3];
 
-            float page_index = 0.0f;
+            uint8_t page_index = 0;
             if (frame_indices != 0x0)
             {
                 uint32_t page_indices_index = frame_indices[tile];
-                page_index                  = (float) page_indices[page_indices_index];
+                page_index                  = (uint8_t) page_indices[page_indices_index];
             }
 
             particle_transform.SetTranslation(Vector3(particle->GetPosition()));
